added a useEngine hook
